+++ crates/ruma-common/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+Breaking changes:
+
+- `OAuthClientScope` now requires that custom values are valid OAuth scopes.
+  Consequently, its `From<str>` implementation has been replaced with a `TryFrom<str>`
+  implementation, and deserializing it is no longer infallible.
+
+Improvements:
+
+- Add unstable support for [MSC4363] "OAuth step up authentication".
+
+[MSC4363]: https://github.com/matrix-org/matrix-spec-proposals/pull/4363
+
 ## 0.20.0
 
 Breaking changes: